Rename package datastructure to model for convention.
