+++ src/main/java/model/VariantInformation.java
@@ -1,4 +1,4 @@
-package datastructure;
+package model;
 
 import utility.Constants;
