+++ src/main/java/model/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Contains all classes that constitute the storage data structures of the MUSIAL application.
  */
-package datastructure;
+package model;
